pyproject.toml
[project]
name = "langchain-learning"
version = "0.1.0"
description = ""
authors = [
    {name = "sahilshukla302003",email = "148991684+sahilshukla302003@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain-openai (>=1.0.1,<2.0.0)",
    "langchain-google-genai (>=3.0.0,<4.0.0)",
     "langchain (>=1.0.2,<2.0.0)",
    "langchain-text-splitters (>=1.0.0,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "chromadb (>=1.2.1,<2.0.0)",
    "sentence-transformers (>=5.1.2,<6.0.0)",
    "langchain-huggingface (>=1.0.0,<2.0.0)",
    "langchain-chroma (>=1.0.0,<2.0.0)",
    "langchain-core (>=1.0.1,<2.0.0)",
    "pypdf (>=6.1.3,<7.0.0)",
    "pymupdf (>=1.26.5,<2.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "langgraph (>=1.0.1,<2.0.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "smolagents[litellm] (>=1.22.0,<2.0.0)",
    "ddgs (>=9.6.1,<10.0.0)",
    "llama-index (>=0.14.7,<0.15.0)",
    "llama-index-core (>=0.14.7,<0.15.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
